Edit questions in README for clarity and consistency
Changes to:
* Formatting
* Capitalization
* Sentence structure, where appropriate
* Question-ifying (please confirm [...]. -> Do you [...]?)

I had a hard time understanding a few of the questions, and spent some
time looking through the history to understand when they were added and
how they evolved. Some of them were phased differently between
ISSUE_TEMPLATE and README, so when in doubt I've erred on the side of
keeping more detailed versions of questions.