Move all questions from ISSUE_TEMPLATE to README
This is a bit of a workflow change. Based on the conversation in
https://github.com/rhboot/shim-review/pull/207, seems like the README
should be the source of truth for submissions.

I've tried to remove duplicates. When in doubt I've used the history to
see what questions were added at the same time and considered
similar-but-different phrasing to be "duplicated".

For now all added questions have been tacked on the end. Grouping by
subject can come later.