Use unambiguous character for horizontal lines
In commonmark, `---` and `===` can be used to mark either [setext
headings] or [thematic breaks] (aka horizontal lines). Headings take
precendence, so if you aren't careful with line breaks you can make a
heading where you meant to have a horizontal line. See [example] for a
case of this happening.

Fortunately, `***` is unambiguous: it will always create a horizontal
line instead of a heading. Switch all the separators to that format so
that we never have to worry about accidental headings again.

[setext headings]: https://spec.commonmark.org/0.30/#setext-headings
[thematic breaks]: https://spec.commonmark.org/0.30/#thematic-breaks
[example]: https://github.com/rhboot/shim-review/blob/b8ebe98d7198174e95d9e62e4522c145ee9caa5b/README.md#this-should-include-logs-for-creating-the-buildroots-applying-patches-doing-the-build-creating-the-archives-etc